modularized emulator scripts.
this was needed because a couple steps in the process are flaky, and it was no good to rerun the whole process. this brings DRY readability improvements as well as efficiency. should be able run ./scripts/start_flow_emulate.sh with an abi (-a, they all work) and get a very easy path to running the app. if things go wrong, simply read start_flow_emulate.sh, there are two nested layers of control. running ./scripts/flow_emulator_run_apk.sh restarts it if closed without repeating the setup work.